+++ cpp/libcore/source/geometry/world_builder.hpp
@@ -1,27 +1,38 @@
 #ifndef SIMULATOR_WORLD_BUILDER_HPP
 #define SIMULATOR_WORLD_BUILDER_HPP
 
-#include "coordinate.hpp"
-#include "line_segment.hpp"
-#include "special_area.hpp"
-
-#include <map>
 #include <vector>
+#include <map>
+
+struct Point{
+    int x;
+    int y;
+};
+
+struct Line{
+    Point start;
+    Point end;
+};
+
+struct Area{
+    int id;
+    std::vector<Line> lines;
+    //TODO: special properties
+};
 
 class WorldBuilder
 {
 private:
-    std::map<int, std::vector<LineSegment>> m_level_to_boundary_lines;
-    std::map<int, std::vector<SpecialArea>> m_level_to_special_areas;
+    std::map<int, std::vector<Line>> m_level_to_boundary_lines;
+    std::map<int, std::vector<Area>> m_level_to_special_areas;
 
 public:
+
     WorldBuilder() = default;
 
-    // TODO: somehow check for maximum dimensions and shift if necessary (to not exceed the int
-    // range)
-
-    void addBoundaryLine(LineSegment p_segment);
-    void addSpecialArea(SpecialArea p_area);
+    //TODO: somehow check for maximum dimensions and shift if necessary (to not exceed the int range)
+    void addBoundaryLine(int p_level, std::vector<int> p_segment);
+    void addSpecialArea(int p_level, int p_id, std::vector<int> p_segments);
 };
 
-#endif // SIMULATOR_WORLD_BUILDER_HPP
+#endif //SIMULATOR_WORLD_BUILDER_HPP
